+++ user-service/src/main/java/com/palja/user_service/application/event/dto/request/DeleteCompanyUserEventReq.java
@@ -1,4 +1,4 @@
-package com.palja.user_service.application.event.dto.impl;
+package com.palja.user_service.application.event.dto.request;
 
 import java.util.UUID;
