+++ user-service/src/main/java/com/palja/user_service/application/service/impl/CustomerServiceImpl.java
@@ -1,6 +1,5 @@
 package com.palja.user_service.application.service.impl;
 
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -16,7 +15,8 @@ import com.palja.user_service.application.dto.response.CreateUserRes;
 import com.palja.user_service.application.dto.response.ReadCustomerDetailRes;
 import com.palja.user_service.application.dto.response.ReadCustomerSummaryRes;
 import com.palja.user_service.application.dto.response.UpdateCustomerDetailRes;
-import com.palja.user_service.application.event.dto.impl.DeleteCustomerEventReq;
+import com.palja.user_service.application.event.dto.request.DeleteCustomerEventReq;
+import com.palja.user_service.application.event.publisher.UserDomainEventPublisher;
 import com.palja.user_service.application.exception.AuthErrorCode;
 import com.palja.user_service.application.exception.UserErrorCode;
 import com.palja.user_service.application.service.CustomerService;
@@ -34,7 +34,7 @@ public class CustomerServiceImpl implements CustomerService {
 	private final UserRepository userRepository;
 	private final TokenRepository tokenRepository;
 
-	private final ApplicationEventPublisher applicationEventPublisher;
+	private final UserDomainEventPublisher userDomainEventPublisher;
 
 	private final PasswordEncoder passwordEncoder;
 	private final JwtUtil jwtUtil;
@@ -117,7 +117,7 @@ public class CustomerServiceImpl implements CustomerService {
 		User user = getCustomerByLoginId(loginId);
 		user.softDelete();
 
-		applicationEventPublisher.publishEvent(DeleteCustomerEventReq.from(user.getId()));
+		userDomainEventPublisher.publishEvent(DeleteCustomerEventReq.from(user.getId()));
 	}
 
 	@Override
@@ -128,7 +128,7 @@ public class CustomerServiceImpl implements CustomerService {
 		User user = getCustomerByLoginId(currentUserLoginId);
 		user.softDelete();
 
-		applicationEventPublisher.publishEvent(DeleteCustomerEventReq.from(user.getId()));
+		userDomainEventPublisher.publishEvent(DeleteCustomerEventReq.from(user.getId()));
 
 		String substringAccessToken = jwtUtil.substringToken(accessToken);
 		String hashKey = jwtUtil.hashingTokenToSHA256(substringAccessToken);
